backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.6
flower==2.0.1

# HTTP клиент
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Парсинг и scraping
feedparser==6.0.10
beautifulsoup4==4.12.3
lxml==5.1.0

# Telegram
python-telegram-bot==20.7

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Валидация и утилиты
email-validator==2.1.0
python-dateutil==2.8.2

# Scheduler
apscheduler==3.10.4

# Logging
structlog==24.1.0

# Retry logic
tenacity==8.2.3

# Environment
python-dotenv==1.0.0

# Мониторинг
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
